+++ packages/orq-rc/RELEASES.md
@@ -7669,3 +7669,13 @@ Based on:
 - [python v4.9.0-rc.34] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.34] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.34 - packages/orq-rc
+
+## 2026-05-08 11:25:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.35] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.35 - packages/orq-rc